[WIP] got all tests passing
