+++ lib/pinchflat/filesystem/filesystem_data_worker.ex
@@ -1,4 +1,4 @@
-defmodule Pinchflat.Workers.FilesystemDataWorker do
+defmodule Pinchflat.Filesystem.FilesystemDataWorker do
   @moduledoc false
 
   use Oban.Worker,
@@ -7,7 +7,7 @@ defmodule Pinchflat.Workers.FilesystemDataWorker do
     max_attempts: 1
 
   alias Pinchflat.Media
-  alias Pinchflat.Tasks.MediaItemTasks
+  alias Pinchflat.Filesystem.FilesystemHelpers
 
   @impl Oban.Worker
   @doc """
@@ -18,7 +18,7 @@ defmodule Pinchflat.Workers.FilesystemDataWorker do
   def perform(%Oban.Job{args: %{"id" => media_item_id}}) do
     media_item = Media.get_media_item!(media_item_id)
 
-    MediaItemTasks.compute_and_save_media_filesize(media_item)
+    FilesystemHelpers.compute_and_save_media_filesize(media_item)
 
     # Don't retry on failure - if it didn't work immediately there's no
     # reason to believe it will work later.
