+++ lib/pinchflat/filesystem/filesystem_helpers.ex
@@ -1,8 +1,8 @@
-defmodule Pinchflat.Utils.FilesystemUtils do
+defmodule Pinchflat.Filesystem.FilesystemHelpers do
   @moduledoc """
   Utility methods for working with the filesystem
   """
-
+  alias Pinchflat.Media
   alias Pinchflat.Utils.StringUtils
 
   @doc """
@@ -19,5 +19,20 @@ defmodule Pinchflat.Utils.FilesystemUtils do
     :ok = File.write(filepath, "")
 
     filepath
+  end
+
+  @doc """
+  Fetches the file size of a media item and saves it to the database.
+
+  Returns {:ok, media_item} | {:error, any()}
+  """
+  def compute_and_save_media_filesize(media_item) do
+    case File.stat(media_item.media_filepath) do
+      {:ok, %{size: size}} ->
+        Media.update_media_item(media_item, %{media_size_bytes: size})
+
+      err ->
+        err
+    end
   end
 end
