--- lib/pinchflat/tasks/source_tasks.ex
@@ -1,25 +1,0 @@
-defmodule Pinchflat.Tasks.SourceTasks do
-  @moduledoc """
-  Contains methods used by OR used to create/manage tasks for sources.
-
-  Tasks/workers are meant to be thin wrappers so most of the actual work they
-  do is also defined here. Essentially, a one-stop-shop for source-related tasks/workers.
-  """
-
-  require Logger
-
-  # alias Pinchflat.Media
-  # alias Pinchflat.Tasks
-  # alias Pinchflat.Sources
-  # alias Pinchflat.Sources.Source
-  # alias Pinchflat.FastIndexing.YoutubeRss
-  # alias Pinchflat.Media.MediaItem
-  # alias Pinchflat.FastIndexing.FastIndexingWorker
-  # alias Pinchflat.Downloading.MediaDownloadWorker
-  # alias Pinchflat.FastIndexing.MediaIndexingWorker
-  # alias Pinchflat.YtDlp.Backend.MediaCollection
-  # alias Pinchflat.SlowIndexing.MediaCollectionIndexingWorker
-  # alias Pinchflat.Utils.FilesystemUtils.FileFollowerServer
-
-  # alias Pinchflat.YtDlp.Backend.Media, as: YtDlpMedia
-end
