--- test/pinchflat/tasks/source_tasks_test.exs
@@ -1,20 +1,0 @@
-defmodule Pinchflat.Tasks.SourceTasksTest do
-  use Pinchflat.DataCase
-
-  import Mox
-  import Pinchflat.TasksFixtures
-  import Pinchflat.MediaFixtures
-  import Pinchflat.SourcesFixtures
-  import Pinchflat.ProfilesFixtures
-
-  alias Pinchflat.Tasks
-  alias Pinchflat.Tasks.Task
-  alias Pinchflat.Tasks.SourceTasks
-  alias Pinchflat.Media.MediaItem
-  alias Pinchflat.FastIndexing.FastIndexingWorker
-  alias Pinchflat.Downloading.MediaDownloadWorker
-  alias Pinchflat.FastIndexing.MediaIndexingWorker
-  alias Pinchflat.SlowIndexing.MediaCollectionIndexingWorker
-
-  setup :verify_on_exit!
-end
